Handle it better when a managed gRPC server crashes on startup
Summary: This should not be very common, but a user did hit it when the dagster.grpc module wasn't available. Instead of crashing and then waiting 15 seconds for a message that never comes, then timing out, detect right away that the process is no longer alive while waiting for the message.

Test Plan: New Test

Reviewers: alangenfeld, sashank, max

Differential Revision: https://dagster.phacility.com/D5166